+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - chapter-5.rb [examples from chapter-5]
 - angry-boss.rb [todo from chapter-5]
+- table-of-contents.rb [todo from chapter-5]
 
 
 ## [0.0.2] (May 9 2021)